requirements.txt
Django==3.2.13
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
markdown==3.4.1
django-filter==22.1
django-redis==5.2.0
django-crontab==0.7.1
scipy==1.7.3
matplotlib==3.5.2
basemap==1.3.3
xarray==0.20.2
pymysql==1.0.2
gevent==21.12.0
netCDF4==1.5.8
pydocx==0.9.10
pymupdf==1.19.6
rasterio==1.2.10
geopandas==0.10.2
fiona==1.8.21
Cartopy==0.20.2
h5py==3.6.0